+++ doc/source/changelog.md
@@ -41,6 +41,7 @@ Godot 4 is required from this version.
     - `VoxelGeneratorGraph`: editor: fix inspector starting to throw errors after deleting a node, as it is still inspecting it
     - `VoxelGeneratorGraph`: editor: fixed crash when connecting an SdfPreview node to an input. However this is not supported yet.
     - `VoxelGeneratorGraph`: fixed Image2D node not accepting image formats L8 and LA8
+    - `VoxelMesherCubes`: editor: color mode is now a proper dropdown
     - `VoxelMesherTransvoxel`: fixed surface not appearing if it lines up exactly at integer coordinates
     - `VoxelTerrain`: fixed `Condition "mesh_block == nullptr" is true` which could happen in some conditions
     - `VoxelTerrain`: changing a material now updates existing meshes instead of only new ones
@@ -65,6 +66,7 @@ Godot 4 is required from this version.
     - `VoxelStreamScript`: renamed `_emerge_block` => `_load_voxel_block`
     - `VoxelStreamScript`: renamed `_immerge_block` => `_save_voxel_block`
     - `VoxelGeneratorGraph`: the `Select` node's `threshold` port is now a parameter instead.
+    - `FastNoiseLite` was renamed `ZN_FastNoiseLite`, as now Godot 4 comes with its own implementation, with a few differences.
 
 - Known issues
     - Some nodes and resources no longer start with predefined properties due to a warning introduced in Godot4 when properties are resources.